Refactored code
